# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Translalia API
  version: 2026-03-12
  description: >
    Current-code-first catalog of the Next.js API routes in `translalia-web/src/app/api`.
    Stable/high-use routes have fuller request and response schemas; lower-signal internal
    routes are cataloged with generic object shapes so they remain discoverable.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/reflection/express-your-view:
    get:
      tags: [reflection]
      summary: Fetch the student's post-AI "Express Your View" reflection
      x-classification: public
      parameters:
        - in: query
          name: threadId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Express Your View reflection (state.express_your_view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
    post:
      tags: [reflection]
      summary: Save the student's "Express Your View" reflection
      x-classification: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericObject"
      responses:
        "200":
          description: Updated reflection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
...
  /api/reflection/ai-assist-step-c:
    post:
      tags: [reflection]
      summary: Reflection-rail contextual suggestions
      x-classification: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericObject"
      responses:
        "200":
          description: Reflection assist response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericObject"
...
components:
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
...